requirements.txt
contourpy==1.2.1
cycler==0.12.1
exceptiongroup==1.2.1
fonttools==4.51.0
imageio==2.34.1
importlib_resources==6.4.0
iniconfig==2.0.0
kiwisolver==1.4.5
lazy_loader==0.4
libsvm-official==3.32.0
matplotlib==3.8.4
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
pyparsing==3.1.2
pytest==8.2.0
python-dateutil==2.9.0.post0
PyWavelets==1.6.0
scikit-image==0.22.0
scipy==1.13.0
setuptools-scm==8.0.4
six==1.16.0
tifffile==2024.4.24
tomli==2.0.1
typing_extensions==4.11.0
zipp==3.18.1
